first of the genesis tests passing, some RLP refactoring

+++ src/Nevermind/Nevermind.Core/HomesteadDifficultyCalculator.cs
@@ -4,7 +4,7 @@ namespace Nevermind.Core
 {
     public class HomesteadDifficultyCalculator : FrontierDifficultyCalculator
     {
-        protected internal override BigInteger TimeAdjustment(BigInteger parentTimestamp, BigInteger currentTimestamp, ulong blockNumber)
+        protected internal override BigInteger TimeAdjustment(BigInteger parentTimestamp, BigInteger currentTimestamp, BigInteger blockNumber)
         {
             BigInteger timeAdjustment = BigInteger.Max(1 - BigInteger.Divide(currentTimestamp - parentTimestamp, 10), -99);
             return timeAdjustment;
