+++ src/Nevermind/Nevermind.Core/OlimpicDifficultyCalculator.cs
@@ -4,7 +4,7 @@ namespace Nevermind.Core
 {
     public class OlimpicDifficultyCalculator : FrontierDifficultyCalculator
     {
-        protected internal override BigInteger TimeAdjustment(BigInteger parentTimestamp, BigInteger currentTimestamp, ulong blockNumber)
+        protected internal override BigInteger TimeAdjustment(BigInteger parentTimestamp, BigInteger currentTimestamp, BigInteger blockNumber)
         {
             BigInteger timeAdjustment = currentTimestamp < parentTimestamp + 7 ? 1 : -1;
             return timeAdjustment;
